split up major parts of test_cloudformation.py and move rest to legacy
