+++ tests/integration/cloudformation/test_cloudformation_cdk.py
@@ -7,7 +7,9 @@ import requests
 from localstack import config
 from localstack.utils.aws import aws_stack
 from localstack.utils.files import load_file
+from localstack.utils.strings import short_uid
 from localstack.utils.sync import wait_until
+from localstack.utils.testutil import create_zip_file
 
 
 class TestCdkInit:
@@ -58,3 +60,36 @@ class TestCdkInit:
             # clean up
             cleanup_changesets([change_set_name])
             cleanup_stacks([stack_name])
+
+    # TODO: remove this and replace with CDK test
+    def test_cdk_template(self, s3_client, lambda_client, deploy_cfn_template, s3_create_bucket):
+        bucket = f"bucket-{short_uid()}"
+        key = f"key-{short_uid()}"
+        path = os.path.join(os.path.dirname(__file__), "../templates/asset")
+
+        s3_create_bucket(Bucket=bucket)
+        s3_client.put_object(Bucket=bucket, Key=key, Body=create_zip_file(path, get_content=True))
+
+        template = load_file(
+            os.path.join(os.path.dirname(__file__), "../templates/cdktemplate.json")
+        )
+
+        stack = deploy_cfn_template(
+            template=template,
+            parameters={
+                "AssetParameters1S3BucketEE4ED9A8": bucket,
+                "AssetParameters1S3VersionKeyE160C88A": key,
+            },
+        )
+
+        resp = lambda_client.list_functions()
+        functions = [func for func in resp["Functions"] if stack.stack_name in func["FunctionName"]]
+
+        assert len(functions) == 2
+        assert (
+            len([func for func in functions if func["Handler"] == "index.createUserHandler"]) == 1
+        )
+        assert (
+            len([func for func in functions if func["Handler"] == "index.authenticateUserHandler"])
+            == 1
+        )
