+++ tests/integration/cloudformation/test_cloudformation_sqs.py
@@ -1,5 +1,10 @@
 import os
 
+import pytest
+from botocore.exceptions import ClientError
+
+from localstack.utils.strings import short_uid
+
 
 def test_sqs_queue_policy(sqs_client, deploy_cfn_template):
     result = deploy_cfn_template(
@@ -32,3 +37,44 @@ def test_sqs_non_fifo_queue_generates_valid_name(deploy_cfn_template):
         template_mapping={"is_fifo": "false"},
     )
     assert ".fifo" not in result.outputs["FooQueueName"]
+
+
+TEST_TEMPLATE_15 = """
+AWSTemplateFormatVersion: 2010-09-09
+Resources:
+  FifoQueue:
+    Type: 'AWS::SQS::Queue'
+    Properties:
+        QueueName: %s
+        ContentBasedDeduplication: "false"
+        FifoQueue: "true"
+  NormalQueue:
+    Type: 'AWS::SQS::Queue'
+    Properties:
+        ReceiveMessageWaitTimeSeconds: 1
+"""
+
+
+def test_cfn_handle_sqs_resource(self, deploy_cfn_template, sqs_client):
+    fifo_queue = f"queue-{short_uid()}.fifo"
+
+    stack = deploy_cfn_template(template=TEST_TEMPLATE_15 % fifo_queue)
+
+    rs = sqs_client.get_queue_url(QueueName=fifo_queue)
+    assert rs["ResponseMetadata"]["HTTPStatusCode"] == 200
+
+    queue_url = rs["QueueUrl"]
+
+    rs = sqs_client.get_queue_attributes(QueueUrl=queue_url, AttributeNames=["All"])
+    attributes = rs["Attributes"]
+    assert "ContentBasedDeduplication" in attributes
+    assert "FifoQueue" in attributes
+    assert attributes["ContentBasedDeduplication"] == "false"
+    assert attributes["FifoQueue"] == "true"
+
+    # clean up
+    stack.destroy()
+
+    with pytest.raises(ClientError) as ctx:
+        sqs_client.get_queue_url(QueueName=fifo_queue)
+    assert ctx.value.response["Error"]["Code"] == "AWS.SimpleQueueService.NonExistentQueue"
