+++ tests/integration/cloudformation/test_cloudformation_stacks.py
@@ -6,8 +6,10 @@ import pytest
 import yaml
 
 from localstack.testing.aws.cloudformation_utils import load_template_file
-from localstack.utils.common import short_uid
+from localstack.utils.aws import aws_stack
+from localstack.utils.files import load_file
 from localstack.utils.generic.wait_utils import wait_until
+from localstack.utils.strings import short_uid
 from localstack.utils.sync import retry
 
 
@@ -324,3 +326,104 @@ def test_stack_update_resources(
     # describe stack resources
     resources = cfn_client.describe_stack_resources(StackName=stack_name)
     snapshot.match("stack_resources", resources)
+
+
+def test_nested_stack(s3_client, cfn_client, deploy_cfn_template, s3_create_bucket):
+    # upload template to S3
+    artifacts_bucket = f"cf-artifacts-{short_uid()}"
+    artifacts_path = "stack.yaml"
+    s3_create_bucket(Bucket=artifacts_bucket, ACL="public-read")
+    s3_client.put_object(
+        Bucket=artifacts_bucket,
+        Key=artifacts_path,
+        Body=load_file(os.path.join(os.path.dirname(__file__), "../templates/template5.yaml")),
+    )
+
+    # deploy template
+    param_value = short_uid()
+    stack_bucket_name = f"test-{param_value}"  # this is the bucket name generated by template5
+
+    deploy_cfn_template(
+        template=load_file(os.path.join(os.path.dirname(__file__), "../templates/template6.yaml"))
+        % (artifacts_bucket, artifacts_path),
+        parameters={"GlobalParam": param_value},
+    )
+
+    # assert that nested resources have been created
+    def assert_bucket_exists():
+        response = s3_client.head_bucket(Bucket=stack_bucket_name)
+        assert 200 == response["ResponseMetadata"]["HTTPStatusCode"]
+
+    retry(assert_bucket_exists)
+
+
+def test_update_stack_with_same_template(cfn_client, deploy_cfn_template):
+    template = load_file(os.path.join(os.path.dirname(__file__), "../templates/fifo_queue.json"))
+    stack = deploy_cfn_template(template=template)
+
+    with pytest.raises(Exception) as ctx:  # TODO: capture proper exception
+        cfn_client.update_stack(StackName=stack.stack_name, TemplateBody=template)
+        cfn_client.get_waiter("stack_update_complete").wait(StackName=stack.stack_name)
+
+    error_message = str(ctx.value)
+    assert "UpdateStack" in error_message
+    assert "No updates are to be performed." in error_message
+
+
+TEST_TEMPLATE_26_1 = """
+AWSTemplateFormatVersion: 2010-09-09
+Resources:
+  MyQueue:
+    Type: 'AWS::SQS::Queue'
+    Properties:
+      QueueName: %s
+Outputs:
+  TestOutput26:
+    Value: !GetAtt MyQueue.Arn
+    Export:
+      Name: TestQueueArn26
+"""
+TEST_TEMPLATE_26_2 = """
+AWSTemplateFormatVersion: 2010-09-09
+Resources:
+  MessageQueue:
+    Type: 'AWS::SQS::Queue'
+    Properties:
+      QueueName: %s
+      RedrivePolicy:
+        deadLetterTargetArn: !ImportValue TestQueueArn26
+        maxReceiveCount: 3
+Outputs:
+  MessageQueueUrl1:
+    Value: !ImportValue TestQueueArn26
+  MessageQueueUrl2:
+    Value: !Ref MessageQueue
+"""
+
+
+# TODO: re-write this
+@pytest.mark.skip(reason="flaky due to issues in parameter handling and re-resolving")
+def test_stack_imports(deploy_cfn_template, cfn_client, sqs_client):
+    result = cfn_client.list_imports(ExportName="_unknown_")
+    assert result["ResponseMetadata"]["HTTPStatusCode"] == 200
+    assert result["Imports"] == []  # TODO: create test with actual import values!
+
+    queue_name1 = f"q-{short_uid()}"
+    queue_name2 = f"q-{short_uid()}"
+    template1 = TEST_TEMPLATE_26_1 % queue_name1
+    template2 = TEST_TEMPLATE_26_2 % queue_name2
+    deploy_cfn_template(template=template1)
+    stack2 = deploy_cfn_template(template=template2)
+
+    queue_url1 = sqs_client.get_queue_url(QueueName=queue_name1)["QueueUrl"]
+    queue_url2 = sqs_client.get_queue_url(QueueName=queue_name2)["QueueUrl"]
+
+    queues = sqs_client.list_queues().get("QueueUrls", [])
+    assert queue_url1 in queues
+    assert queue_url2 in queues
+
+    outputs = cfn_client.describe_stacks(StackName=stack2.stack_name)["Stacks"][0]["Outputs"]
+    output = [out["OutputValue"] for out in outputs if out["OutputKey"] == "MessageQueueUrl1"][0]
+    assert aws_stack.sqs_queue_arn(queue_url1) == output  # TODO
+    output = [out["OutputValue"] for out in outputs if out["OutputKey"] == "MessageQueueUrl2"][0]
+    assert output == queue_url2
