--- tests/integration/templates/template21.json
@@ -1,55 +1,0 @@
-{
-    "AWSTemplateFormatVersion": "2010-09-09",
-    "Transform": "AWS::Serverless-2016-10-31",
-    "Resources": {
-        "DynamoOnDemand": {
-            "Type": "AWS::DynamoDB::Table",
-            "Properties": {
-                "TableName": "ondemandtable",
-                "AttributeDefinitions": [
-                    {
-                        "AttributeName": "foo",
-                        "AttributeType": "S"
-                    },
-                    {
-                        "AttributeName": "bar",
-                        "AttributeType": "S"
-                    },
-                    {
-                        "AttributeName": "baz",
-                        "AttributeType": "S"
-                    }
-                ],
-                "KeySchema": [
-                    {
-                        "AttributeName": "foo",
-                        "KeyType": "HASH"
-                    },
-                    {
-                        "AttributeName": "bar",
-                        "KeyType": "RANGE"
-                    }
-                ],
-                "GlobalSecondaryIndexes": [
-                    {
-                        "IndexName": "by.baz",
-                        "KeySchema": [
-                            {
-                                "AttributeName": "foo",
-                                "KeyType": "HASH"
-                            },
-                            {
-                                "AttributeName": "baz",
-                                "KeyType": "RANGE"
-                            }
-                        ],
-                        "Projection": {
-                            "ProjectionType": "ALL"
-                        }
-                    }
-                ],
-                "BillingMode": "PAY_PER_REQUEST"
-            }
-        }
-    }
-}
